fix: parse packets from right data source
The bug is that the same packet will be parsed for as many times as the
size of that packet would fit in the statically allocated `data` buffer.

Could we have more easily spotted this error? Perhaps with a better name
for `data`? Such as `s_data` to indicate that it is statically
allocated?